응답 모델의 일부 데이터 타입 변경

+++ app/src/main/java/com/bookfinder/model/Book.kt
@@ -24,7 +24,7 @@ object Book {
                 var pageCount: Int,
                 var printType: String,
                 var categories: ArrayList<String>,
-                var averageRating: Int,
+                var averageRating: Float,
                 var ratingsCount: Int,
                 var maturityRating: String,
                 var allowAnonLogging: Boolean,
